exitQuiz: ändrade metod

+++ src/main/java/com/eq/house/PlayController.java
@@ -388,9 +388,6 @@ public class PlayController {
 		if (answers == null) {
 			model.addAttribute("errorMsg", "exitQuiz: fatal: cannot find answers");
 			return "error";
-		} else if (answers.isEmpty()) {
-			model.addAttribute("errorMsg", "exitQuiz: fatal: zero answers");
-			return "error";
 		} else if ((quiz = getQuizByUniqueId(quizId)) == null) {
 			model.addAttribute("errorMsg", "exitQuiz: fatal: no such quiz");
 			return "error";
@@ -399,7 +396,8 @@ public class PlayController {
 		System.out.println("----- exitQuiz: -----");
 		System.out.println("answers:    " + answers.size());
 		System.out.println("pre count:  " + answerRepo.count());
-		answerRepo.deleteByPlayerId(playerId);
+		if (!answers.isEmpty())
+			answerRepo.deleteByPlayerId(playerId);
 		System.out.println("post count: " + answerRepo.count());
 
 		model.addAttribute("quiz", quiz);
